Fix: Corrigir tipo FileData para Blob no videoCompressor
- Adicionar type assertion para Uint8Array
- Resolver erro de tipo FileData não compatível com BlobPart

+++ lib/videoCompressor.ts
@@ -178,8 +178,9 @@ export async function compressVideo(
     await ffmpeg.deleteFile(inputFileName);
     await ffmpeg.deleteFile(outputFileName);
 
-    // Converter para Blob
-    const compressedBlob = new Blob([data], { type: 'video/mp4' });
+    // Converter FileData (Uint8Array) para Blob
+    // readFile retorna Uint8Array, que precisa ser convertido corretamente
+    const compressedBlob = new Blob([data as Uint8Array], { type: 'video/mp4' });
     const compressedSize = compressedBlob.size;
     const reduction = ((originalSize - compressedSize) / originalSize * 100).toFixed(1);
     console.log(`Compressão concluída: ${file.name} - ${reduction}% menor (${(originalSize / 1024 / 1024).toFixed(2)} MB → ${(compressedSize / 1024 / 1024).toFixed(2)} MB)`);
